have just a single SubmitCog that dispatches submissions to the correct launcher
